+++ test/test/define.h
@@ -1,9 +1,11 @@
 #pragma once
 
-double Road_Longitude;
-double Road_latitude;
-int    Region_ID;
-int    Limit_Speed;
-double Signage_Longitude;
-double Signage_Latitude;
+float Road_Longitude;
+float Road_latitude;
+int   Region_ID;
+int   Limit_Speed;
+
+
+float Signage_Longitude;
+float Signage_Latitude;
 int    Warn_Type;
